requirements.txt
# Core ML/DL
torch>=2.1.0
torchaudio>=2.1.0
transformers>=4.35.0
numpy>=1.24.0
scipy>=1.11.0

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1
noisereduce>=3.0.0
pyannote.audio>=3.1.0
speechbrain>=0.5.16
whisperx>=3.1.1

# NLP & LLM
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-google-genai>=0.0.5
openai>=1.3.0
tiktoken>=0.5.1
chromadb>=0.4.18
faiss-cpu>=1.7.4

# API & Web
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1
celery>=5.3.4
redis>=5.0.1

# Frontend
streamlit>=1.28.0
plotly>=5.18.0
streamlit-webrtc>=0.47.0

# MLOps & Monitoring
wandb>=0.16.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dateutil>=2.8.2
python-json-logger>=2.0.7
tenacity>=8.2.3
requests>=2.31.0

